Added restore method to restore the whole SDK
So all services can be restored at once, without it being necessary to pay attention to what’s been stubbed, this change adds an array to store the list of services for which methods have been stubbed; this list is populated when the service is stubbed.

The restore method then iterates over the array, restoring each service (and removing it from the array) in turn.